[Controller_Motor_STM32] Added a function to stop following the current trajectory.
This function can be used to stop following the current trajectory and delete
other queued trajectories.
The code has been added in the control program to use this ability when the
stop button is pressed in the viewer.

(note that the real robot will brake with a certain acceleration but that the
simulator will instantly stop.)